+++ src/app/services/storageServices.js
@@ -1,0 +1,5 @@
+
+export const saveLastInterestProduct = (product) => {
+  //?API local storage / SessionStorage
+  localStorage.setItem("lastProduct", product);
+};
