+++ src/index.css
@@ -12,3 +12,8 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+img{
+  max-height: 30vh;
+  max-width: 100%;
+}
